Updates after organization rename
