+++ aap-db-deploy-quarkus/src/main/resources/aap-db-deploy.sw.yaml
@@ -12,7 +12,7 @@ start: LaunchSoftwareTemplate
 extensions:
   - extensionid: workflow-uri-definitions
     definitions:
-      notifications: "https://raw.githubusercontent.com/parodos-dev/serverless-workflows/main/workflows/shared/specs/notifications-openapi.yaml"
+      notifications: "https://raw.githubusercontent.com/rhdhorchestrator/serverless-workflows/main/workflows/shared/specs/notifications-openapi.yaml"
 functions:
   - name: LaunchSoftwareTemplate
     operation: specs/scaffolder-openapi.yaml#createTask
