+++ aap-job/src/main/resources/aap-job.sw.yaml
@@ -10,7 +10,7 @@ start: LaunchJob
 extensions:
   - extensionid: workflow-uri-definitions
     definitions:
-      notifications: https://raw.githubusercontent.com/parodos-dev/serverless-workflows/main/shared/specs/notifications-openapi.yaml
+      notifications: https://raw.githubusercontent.com/rhdhorchestrator/serverless-workflows/main/shared/specs/notifications-openapi.yaml
 functions:
   - name: launchJob
     operation: specs/aap-openapi.yaml#launchJob
