+++ assessment/assessment-with-custom-code/pom.xml
@@ -4,9 +4,9 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.redhat.rhdh.orchestrator</groupId>
   <artifactId>serverless-workflow-assessment-with-custom-code</artifactId>
-  
+
   <parent>
-    <groupId>dev.parodos</groupId>
+    <groupId>io.rhdhorchestrator</groupId>
     <artifactId>swf-parent</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <relativePath>../../swf-parent</relativePath>
