+++ assessment/assessment-with-kn-function/pom.xml
@@ -4,9 +4,9 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.redhat.rhdh.orchestrator</groupId>
   <artifactId>serverless-workflow-assessment-with-kn-function</artifactId>
-  
+
   <parent>
-    <groupId>dev.parodos</groupId>
+    <groupId>io.rhdhorchestrator</groupId>
     <artifactId>swf-parent</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <relativePath>../../swf-parent</relativePath>
@@ -61,7 +61,7 @@
       <groupId>org.kie.kogito</groupId>
       <artifactId>kogito-addons-quarkus-fabric8-kubernetes-service-catalog</artifactId>
     </dependency>
-    
+
     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-junit5</artifactId>
