+++ cluster-onboarding/cluster-onboarding-func/pom.xml
@@ -4,7 +4,7 @@
   <modelVersion>4.0.0</modelVersion>
   <artifactId>cluster-onboarding-func</artifactId>
   <parent>
-    <groupId>dev.parodos</groupId>
+    <groupId>io.rhdhorchestrator</groupId>
     <artifactId>swf-parent</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <relativePath>../../swf-parent</relativePath>
