+++ cluster-onboarding/cluster-onboarding-func/src/main/java/io/rhdhorchestrator/CloudEventsCustomizer.java
@@ -1,4 +1,4 @@
-package dev.parodos;
+package io.rhdhorchestrator;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
