+++ cluster-onboarding/cluster-onboarding-func/src/main/resources/META-INF/resources/index.html
@@ -235,7 +235,7 @@
                         <div class="app-details">
                             <h5>Application</h5>
                             <ul>
-                                <li>GroupId: <code>dev.parodos</code></li>
+                                <li>GroupId: <code>io.rhdhorchestrator</code></li>
                                 <li>ArtifactId: <code>clusteronboarding-kfunc</code></li>
                                 <li>Version: <code>1.0.0-SNAPSHOT</code></li>
                                 <li>Quarkus Version: <code>3.4.1</code></li>
