+++ cluster-onboarding/serverless-workflow-clusteronboarding/src/main/java/com/redhat/rhdhorchestrator/workflow/clusteronboarding/FileReader.java
@@ -1,4 +1,4 @@
-package com.redhat.parodos.workflow.clusteronboarding;
+package com.redhat.rhdhorchestrator.workflow.clusteronboarding;
 
 
 import jakarta.enterprise.context.ApplicationScoped;
