+++ cluster-onboarding/serverless-workflow-clusteronboarding/src/main/resources/clusteronboarding.sw.yaml
@@ -17,7 +17,7 @@ functions:
     operation: "specs/kube.yaml#readCoreV1NamespacedSecret"
   - name: getServiceAccountInfoFunc
     type: custom
-    operation: service:java:com.redhat.parodos.workflow.clusteronboarding.FileReader::readValue
+    operation: service:java:com.redhat.rhdhorchestrator.workflow.clusteronboarding.FileReader::readValue
   - name: sendCloudEvent
     type: custom
     operation: rest:post:http://broker-ingress.knative-eventing.svc.cluster.local/cluster-onboarding/default
@@ -69,7 +69,7 @@ states:
       name: ClusterOnboardAction
       eventRef:
         triggerEventRef: sendEvent
-        data: 
+        data:
           "{
             workflowCallerId: $WORKFLOW.instanceId,
             clusterApiUrl: .clusterAPIUrl,
