+++ escalation-service-now/pom.xml
@@ -6,11 +6,11 @@
   <modelVersion>4.0.0</modelVersion>
   <artifactId>servicenow-escalation</artifactId>
   <parent>
-    <groupId>dev.parodos</groupId>
+    <groupId>io.rhdhorchestrator</groupId>
     <artifactId>swf-parent</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <relativePath>../swf-parent</relativePath>
-  </parent>  
+  </parent>
   <version>1.0.0-SNAPSHOT</version>
   <properties>
     <version.io.quarkiverse.embedded.postgresql>0.1.3</version.io.quarkiverse.embedded.postgresql>
