+++ escalation-service-now/src/main/resources/serviceNowEscalation.sw.yaml
@@ -10,7 +10,7 @@ start: CreateChangeRequest
 extensions:
   - extensionid: workflow-uri-definitions
     definitions:
-      notifications: https://raw.githubusercontent.com/parodos-dev/serverless-workflows/main/shared/specs/notifications-openapi.yaml
+      notifications: https://raw.githubusercontent.com/rhdhorchestrator/serverless-workflows/main/shared/specs/notifications-openapi.yaml
 dataInputSchema: schemas/servicenow-escalation-schema.json
 functions:
   - name: createChangeRequest
@@ -49,7 +49,7 @@ states:
             message: "\"Sleeping before checking the change request \\(.createdChangeRequest.result)\""
         sleep:
           after: PT30S
-      - name: "getChangeRequest" 
+      - name: "getChangeRequest"
         functionRef:
           refName: getChangeRequest
           arguments:
