+++ service-now-aap-job/workflow.sw.yaml
@@ -10,7 +10,7 @@ start: CreateChangeRequest
 extensions:
   - extensionid: workflow-uri-definitions
     definitions:
-      notifications: https://raw.githubusercontent.com/parodos-dev/serverless-workflows/main/shared/specs/notifications-openapi.yaml
+      notifications: https://raw.githubusercontent.com/rhdhorchestrator/serverless-workflows/main/shared/specs/notifications-openapi.yaml
 dataInputSchema: schemas/service-now-app-job-input-schema.json
 functions:
   - name: createChangeRequest
@@ -75,7 +75,7 @@ states:
             message: "\"Sleeping before checking the change request \\(.serviceNow.createdChangeRequest.result)\""
         sleep:
           after: PT10S
-      - name: "getChangeRequest" 
+      - name: "getChangeRequest"
         functionRef:
           refName: getChangeRequest
           arguments:
