+++ tekton/pom.xml
@@ -4,7 +4,7 @@
   xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>dev.parodos</groupId>
+  <groupId>io.rhdhorchestrator</groupId>
   <artifactId>tekton</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <properties>
